Moved the rest stuff for single entries to the repository
