+++ lib/cacheEntryRepository.js
@@ -1,6 +1,5 @@
 'use strict';
-// TODO: should be instance! cannot use Nocca reference
-var $q = require('q');
+var $utils = require('./utils');
 
 module.exports = CacheEntries;
 
@@ -13,6 +12,8 @@ function CacheEntries (Nocca) {
     this.addRecording = addSingleRecording;
     this.exportRecordings = exportRecordings;
     this.name = function() { return 'memory-entries'; };
+    
+    this.init = initRestRoutes;
 
 // --- Simple recording playback
     
@@ -55,5 +56,68 @@ function CacheEntries (Nocca) {
         return (typeof endpoint !== 'undefined' && typeof key !== 'undefined') ? recordings[endpoint][key] : recordings ;
 
     }
+
+// --- REST control methods
+
+    function initRestRoutes() {
+
+        Nocca.pubsub.publish(Nocca.constants.PUBSUB_REST_ROUTE_ADDED, ['GET:/caches', getCaches]);
+        Nocca.pubsub.publish(Nocca.constants.PUBSUB_REST_ROUTE_ADDED, ['PUT:/caches/memory-caches/:endpoint/:entryHash', true, saveOrReplaceCacheEntry]);
+        Nocca.pubsub.publish(Nocca.constants.PUBSUB_REST_ROUTE_ADDED, ['POST:/caches/package', addCachePackage]);
+
+    }
+
+    function getCaches (req, res, config, matches, writeHead, writeEnd) {
+        writeHead(res, 200).writeEnd(JSON.stringify(exportRecordings(), null, 4));
+    }
+
+    function addCachePackage (req, res, config, matches, writeHead, writeEnd) {
+        $utils.readBody(req).then(function(body) {
+
+            body = JSON.parse(body);
+
+            var recordings = exportRecordings();
+
+            // extract from recordings
+            var downloadObj = {};
+
+            if (typeof body.requestKeys !== 'undefined') {
+                body.requestKeys.forEach(function (value) {
+                    downloadObj[value] = recordings[value];
+                });
+            }
+            else {
+                // if no keys specified just download all recorded
+                downloadObj = recordings;
+            }
+
+            writeHead(res, 200, {
+                'Content-Type': 'application/json'
+            }).writeEnd(JSON.stringify(downloadObj));
+
+        }).fail(function() {
+
+            writeHead(res, 400).writeEnd('Request body could not be parsed, is it a valid JSON string?');
+
+        });
+    }
+    
+    function saveOrReplaceCacheEntry(req, res, config, matches, writeHead, writeEnd) {
+        $utils.readBody(req).then(function(body) {
+            
+            body = JSON.parse(body);
+            
+            
+            
+        }).fail(function() {
+            
+            writeHead(res, 400).writeEnd('Request body could not be parsed, is it a valid JSON string?');
+            
+        });
+    }
+
+
+
+
 
 }
