+++ lib/httpApi.js
@@ -32,8 +32,6 @@ function HttpApi (Nocca) {
     
     // Configure routes
     addRoute('GET:/config', getConfig);
-    addRoute('GET:/caches', getCaches);
-    addRoute('POST:/caches/package', addCachePackage);
     addRoute('GET:/enums/scenarios/type', getEnumScenariosType);
     addRoute('GET:/enums/scenarios/repeatable', getEnumScenariosRepeatable);
 
@@ -126,41 +124,6 @@ function HttpApi (Nocca) {
 
     function getConfig (req, res, config) {
         _writeHead(res, 200).writeEnd(JSON.stringify(config, null, 4));
-    }
-
-    function getCaches (req, res, config) {
-        _writeHead(res, 200).writeEnd(JSON.stringify(Nocca.cacheentries.exportRecordings(), null, 4));
-    }
-
-    function addCachePackage (req, res, config) {
-        $utils.readBody(req).then(function(body) {
-
-            body = JSON.parse(body);
-
-            var recordings = Nocca.cacheentries.exportRecordings();
-
-            // extract from recordings
-            var downloadObj = {};
-
-            if (typeof body.requestKeys !== 'undefined') {
-                body.requestKeys.forEach(function (value) {
-                    downloadObj[value] = recordings[value];
-                });
-            }
-            else {
-                // if no keys specified just download all recorded
-                downloadObj = recordings;
-            }
-
-            _writeHead(res, 200, {
-                'Content-Type': 'application/json'
-            }).writeEnd(JSON.stringify(downloadObj));
-
-        }).fail(function() {
-
-            _writeHead(res, 400).writeEnd('Request body could not be parsed, is it a valid JSON string?');
-
-        });
     }
 
     function getEnumScenariosType (req, res) {
